Fixed bug in cummulative use of unitary_propagator
Now it is possible to break down the simulation into smaller intervals,
and simply accumulate the propagators for each of the intervals. The bug
was due to evolution not covering the entire interval -- either the last
step was missed, or part of the last step was missed, which was
particularly troublesome when the simulation was broken down into many
smaller intervals.